Created an InterfaceBase + Authored the Code
* Created an IVariantEntityBase to ensure all important code is encapsulated
* Authored all code
