+++ NeoForge/src/main/java/software/bluelib/exception/CouldNotLoadJSON.java
@@ -2,11 +2,10 @@
 
 package software.bluelib.exception;
 
-import software.bluelib.entity.variant.VariantParameter;
-
 /**
  * A {@code Class} representing an exception that is thrown when a JSON file could not be loaded.
  * This exception provides additional context by including the {@code getResourceId} of the JSON file that failed to load.
+ * @author MeAlam
  */
 public class CouldNotLoadJSON extends RuntimeException {
 
@@ -20,6 +19,7 @@ public class CouldNotLoadJSON extends RuntimeException {
      *
      * @param pMessage {@link String} - The detail message explaining the reason for the exception.
      * @param pResourceId {@link String} - The ID of the resource that could not be loaded.
+     * @author MeAlam
      */
     public CouldNotLoadJSON(String pMessage, String pResourceId) {
         super(pMessage);
@@ -30,6 +30,7 @@ public class CouldNotLoadJSON extends RuntimeException {
      * A {@link String} that retrieves the resource ID of the JSON file that could not be loaded.
      *
      * @return The resource ID as a string.
+     * @author MeAlam
      */
     public String getResourceId() {
         return resourceId;
